requirements.txt
Django==1.6.8
MySQL-python==1.2.5
Pillow==2.6.1
South==1.0.1
argparse==1.2.1
dj-database-url==0.3.0
django-classy-tags==0.5.1
django-cms==3.0.7
django-mptt==0.6.1
django-reversion==1.8.5
django-sekizai==0.7
djangocms-admin-style==0.2.3
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.5.4
djangocms-link==1.5
djangocms-picture==0.1
djangocms-style==1.5
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.1
djangocms-video==0.1
html5lib==0.999
pytz==2014.10
six==1.8.0
virtualenv==1.11.6
wsgiref==0.1.2
